refactor: mutation 공통 에러 로직을 defaultOption으로 분리 #358
