+++ frontend/src/components/StartButtonContainer/StartButton/hooks/useGameStart.ts
@@ -3,32 +3,14 @@ import { useParams } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 
 import { startGame } from '@/apis/room';
-import AlertModal from '@/components/common/AlertModal/AlertModal';
-import useModal from '@/hooks/useModal';
-import useToast from '@/hooks/useToast';
 import { memberInfoState } from '@/recoil/atom';
-import { CustomError, NetworkError } from '@/utils/error';
-
-const isServerError = (status: number) => status >= 500 && status !== 555;
 
 export const useGameStart = () => {
   const [memberInfo, setMemberInfo] = useRecoilState(memberInfoState);
   const { roomId } = useParams();
-  const { show } = useToast();
-  const { show: showModal } = useModal();
 
   const startGameMutation = useMutation({
     mutationFn: () => startGame(Number(roomId)),
-    onError: (error: CustomError) => {
-      if (error instanceof NetworkError) {
-        show(error.message);
-        return;
-      }
-
-      showModal(AlertModal, { title: '게임 시작 에러', message: error.message });
-    },
-    networkMode: 'always',
-    throwOnError: (error: CustomError) => isServerError(error.status),
   });
 
   const handleGameStart = () => {
