+++ frontend/src/components/common/SelectButton/SelectButton.hook.ts
@@ -2,15 +2,8 @@ import { useMutation } from '@tanstack/react-query';
 import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 
-import AlertModal from '../AlertModal/AlertModal';
-
 import { voteBalanceContent } from '@/apis/balanceContent';
-import useModal from '@/hooks/useModal';
-import useToast from '@/hooks/useToast';
 import { memberInfoState } from '@/recoil/atom';
-import { CustomError, NetworkError } from '@/utils/error';
-
-const isServerError = (status: number) => status >= 500 && status !== 555;
 
 interface UseSelectCompleteMutationProps {
   selectedId: number;
@@ -25,8 +18,6 @@ const useCompleteSelectionMutation = ({
 }: UseSelectCompleteMutationProps) => {
   const { roomId } = useParams();
   const memberInfo = useRecoilValue(memberInfoState);
-  const { show } = useToast();
-  const { show: showModal } = useModal();
 
   return useMutation({
     mutationFn: async () => {
@@ -44,16 +35,6 @@ const useCompleteSelectionMutation = ({
     onSuccess: () => {
       completeSelection();
     },
-    onError: (error: CustomError) => {
-      if (error instanceof NetworkError) {
-        show(error.message);
-        return;
-      }
-
-      showModal(AlertModal, { title: '선택 에러', message: error.message });
-    },
-    networkMode: 'always',
-    throwOnError: (error: CustomError) => isServerError(error.status),
   });
 };
