+++ frontend/src/pages/NicknamePage/useMakeOrEnterRoom.ts
@@ -4,9 +4,7 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 
 import { enterRoom, createRoom } from '@/apis/room';
-import AlertModal from '@/components/common/AlertModal/AlertModal';
 import { ROUTES } from '@/constants/routes';
-import useModal from '@/hooks/useModal';
 import { memberInfoState, roomUuidState } from '@/recoil/atom';
 import { CreateOrEnterRoomResponse } from '@/types/room';
 import { CustomError } from '@/utils/error';
@@ -18,7 +16,6 @@ const useMakeOrEnterRoom = () => {
 
   const [, setRoomUuidState] = useRecoilState(roomUuidState);
   const { roomUuid } = useParams();
-  const { show: showModal } = useModal();
 
   const createRoomMutation = useMutation<CreateOrEnterRoomResponse, CustomError, string>({
     mutationFn: createRoom,
@@ -29,9 +26,6 @@ const useMakeOrEnterRoom = () => {
       }));
       setRoomUuidState(data.roomUuid || '');
       navigate(ROUTES.ready(Number(data.roomId)), { replace: true });
-    },
-    onError: (error) => {
-      showModal(AlertModal, { title: '방 생성 에러', message: error.message });
     },
   });
 
@@ -45,9 +39,6 @@ const useMakeOrEnterRoom = () => {
       setMemberInfo((prev) => ({ ...prev, memberId: data.member.memberId }));
       setRoomUuidState(data.roomUuid || '');
       navigate(ROUTES.ready(Number(data.roomId)), { replace: true });
-    },
-    onError: (error: CustomError) => {
-      showModal(AlertModal, { title: '방 참여 에러', message: error.message });
     },
   });
